Add play game and result activity

+++ App/app/src/main/java/com/pfiev/englishcontest/model/AnswerItem.java
@@ -1,34 +1,34 @@
 package com.pfiev.englishcontest.model;
 
 public class AnswerItem {
-    private int question_id;
-    private int choice_id;
+    private long question_id;
+    private long choice_id;
     private int time_answer;
     private boolean is_right;
 
     public AnswerItem() {
     }
 
-    public AnswerItem(int question_id, int choice_id, int time_answer, boolean is_right) {
+    public AnswerItem(long question_id, long choice_id, int time_answer, boolean is_right) {
         this.question_id = question_id;
         this.choice_id = choice_id;
         this.time_answer = time_answer;
         this.is_right = is_right;
     }
 
-    public int getQuestion_id() {
+    public long getQuestion_id() {
         return question_id;
     }
 
-    public void setQuestion_id(int question_id) {
+    public void setQuestion_id(long question_id) {
         this.question_id = question_id;
     }
 
-    public int getChoice_id() {
+    public long getChoice_id() {
         return choice_id;
     }
 
-    public void setChoice_id(int choice_id) {
+    public void setChoice_id(long choice_id) {
         this.choice_id = choice_id;
     }
